Move the device list functionality to the new home view model so that the main view model can work only as a n orchestrator betweem vms
