+++ NetStalkerAvalonia.Core/Views/Components/ListControls.axaml.cs
@@ -10,9 +10,4 @@ public partial class ListControls : UserControl
     {
         InitializeComponent();
     }
-
-    private void InitializeComponent()
-    {
-        AvaloniaXamlLoader.Load(this);
-    }
 }
