+++ NetStalkerAvalonia.Core/Views/Components/StatsBox.axaml
@@ -2,12 +2,12 @@
 			 xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
 			 xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
 			 xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-			 xmlns:viewModels="clr-namespace:NetStalkerAvalonia.Core.ViewModels"
+			 xmlns:viewModels="clr-namespace:NetStalkerAvalonia.Core.ViewModels.RoutedViewModels"
 			 mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
 			 x:Class="NetStalkerAvalonia.Core.Views.Components.StatsBox">
 
 	<Design.DataContext>
-		<viewModels:MainViewModel />
+		<viewModels:HomeViewModel />
 	</Design.DataContext>
 
 	<StackPanel Orientation="Horizontal" Spacing="10">
