+++ NetStalkerAvalonia.Core/Views/Components/StatsBox.axaml.cs
@@ -10,9 +10,4 @@ public partial class StatsBox : UserControl
     {
         InitializeComponent();
     }
-
-    private void InitializeComponent()
-    {
-        AvaloniaXamlLoader.Load(this);
-    }
 }
